update depth map function

+++ utils/visualizers.py
@@ -73,7 +73,8 @@ class Visualizers(object):
         #ext = 'ext'
         #p = os.path.join(path,'test.exr')
         p = os.path.join(static_path,'pred_depth.exr')
-        imageio.imwrite(p, (depth.cpu().numpy())[0, :, :, :].transpose(1,2,0))
+        if os.path.exists(p):
+            imageio.imwrite(p, (depth.cpu().numpy())[0, :, :, :].transpose(1,2,0))
         depth = self._minmax_normalization(depth)
         turbo = functools.partial(self._matplotlib_colormap, cm.get_cmap('turbo'))
         #b, _, __, ___ = depth.shape
